// apps/web/src/api/generated/WorkflowNodeDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { WorkflowNodeInputDto } from "./WorkflowNodeInputDto";

export type WorkflowNodeDto = { id: string, operator: string, stage: string | null, inputs: { [key in string]?: WorkflowNodeInputDto }, config: Record<string, unknown> | null, };